tests/wpt/web-platform-tests/css/css-typed-om/stylevalue-subclasses/numeric-objects/cssUnitValue.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>CSSUnitValue tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#simple-numeric">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<script>
'use strict';

const gInvalidTestUnits = [
  { unit: '', desc: 'an empty string' },
  { unit: 'lemon', desc: 'an invalid unit' },
];

for (const {unit, desc} of gInvalidTestUnits) {
  test(() => {
    assert_throws(new TypeError(), () => new CSSUnitValue(0, unit));
  }, 'Constructing CSSUnitValue with ' + desc + ' as the unit throws a TypeError');
}

for (const unit of gValidUnits) {
  test(() => {
    const result = new CSSUnitValue(-3.14, unit);
    assert_equals(result.value, -3.14);
    assert_equals(result.unit, unit.toLowerCase());
  }, 'CSSUnitValue can be constructed with ' + unit);
}

</script>
